requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework>=3.9,<4.0

django-ordered-model>=3.1,<3.2

flake8>=3.6.0, <3.7.0
